iommu: Ensure device has the same iommu_ops as the domain
The core code should not call an iommu driver op with a struct
device parameter unless it knows that the dev_iommu_priv_get()
for that struct device was setup by the same driver. Otherwise
in a mixed driver system the iommu_priv could be casted to the
wrong type.

Store the iommu_ops pointer in the iommu_domain and use it as
a check to validate that the struct device is correct before
invoking any domain op that accepts a struct device.

It allows removing the check of the domain op equality in VFIO
and expanding in the future to actually check every domain op.

Change-Id: Ib061052b66448eef64dafd6364b636e341d9cd22
Co-developed-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>