iommu: Add new interface to get iommu hardware information
Introduce a new iommu op and an iommu interface to get the IOMMU hardware
capabilities through iommufd. This information will be used by any vIOMMU
driver which is owned by userspace.

This API chooses to make the special parameters opaque to the core. This
suits the current usage model where accessing any of the IOMMU device
special parameters does require a userspace driver that matches the kernel
driver. If a need for common parameters, implemented similarly by several
drivers, arises then there is room in the design to grow a generic paramter
set as well.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>