iommu: Add iommu_get_msi_domain helper
The new IOMMUFD designs nesting setup using two iommu_domain pointers,
corresponding to stage S1 and stage S2, and the S1 iommu_domain has a
type of IOMMU_DOMAIN_NESTING, is a virtual object that represents the
iommu_domain in the guest, which it attaches the device onto. And the
S2 domain is owned by a hypervisor, handling stage-2 IO page table.

So the msi_cookie is actually setup in the s2 iommu_domain. This means
iommu_get_domain_for_dev() no longer works in iommu_dma_get_msi_page()
and iommu_dma_compose_msi_msg(), because that returns s1 iommu_domain.

Note that not all drivers would need this msi_domain. So adding it to
iommu_group would waste memory. Add a helper function for a driver to
return a correct iommu_domain.

Change-Id: Ia132d9cb179433f5b6f3d1e46b78404a3774af3c
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>