iommufd: Add allowed_itree and its add/remove helpers
The allowed_itree is used for reservations of iova ranges for a future
mapping without an IOMMU_IOAS_MAP_FIXED_IOVA flag, i.e. when user space
wants kernel to manage IOVA. This means that any new kernel-managed IOVA
allocation should allocate from this itree, if it's present. And this
allowed_itree must not intersect with the reserved_itree when adding an
IOVA range, vice versa.

Add a pair of new helpers for new ioctl to add/remove nodes in the itree.
Change the iova allocator to use this tree instead of the reserved_itree
since the reserved_itree is a subset of the new itree now. Also enforce
the reserved_itree to check an IOVA range to prevent it from expanding
across any intersection with the allowed_itree.

Change-Id: I680e2736a05d232ea52c64e6142dbdae833a97f5
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>