iommufd: Register fault handler directly
The iopf framework is mainly designed for a bottom half workqueue
since a page response could take a long cycle to finish, so doing
everything in the top half of the ISR is not an optimal solution.
However, the fault handler in IOMMUFD handles an IO Page Fault in
an asynchronized way, simply forwarding an iommu_fault to the user
space. So, it doesn't really matter whether having such a handler
in the top half.

Register the fault handler without letting it go through the IOPF
framework. This also allows to support the non recoverable type of
iommu_fault.

Change-Id: I7d740a188ceaebe4e565650dd6d4c88dd8707c13
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>