iommu: Pass in dev pointer to fault handler
The fault handler is registered and invoked per-device. And the
iommu_queue_iopf is actually bound to the device also. So, it
makes sense to pass in the dev pointer directly.

This allows to load some other private data using the "cookie".
In one of the following patches, IOMMUFD registers the fault
handler for nested tranlsation support, passing in the nested
IO page table info using "cookie".

Change-Id: I8bff837ffffe17d0fefbc6933dd2cefd9504870b
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>