cover-letter: Connect VFIO to IOMMUFD
This series provides an alternative container layer for VFIO implemented
using iommufd. This is optional, if CONFIG_IOMMUFD is not set then it will
not be compiled in.

At this point iommufd can be injected by passing in a iommfd FD to
VFIO_GROUP_SET_CONTAINER which will use the VFIO compat layer in iommufd
to obtain the compat IOAS and the connect up all the VFIO drivers as
appropriate.

This is temporary stopping point, a following series will provide a way to
directly open a VFIO device FD and directly connect it to IOMMUFD. The new
interface will suport more functionality like vPASID and the dynamic
attachment to support nested translation.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>