iommu/vt-d: WA for domain->max_addr check
the max_addr check against s2_domain->max_addr is not addressed yet.
domain->max_addr is set in map/unmap not in the domain alloc. So before
there is any mapping to this domain, this check will fail. This patch
just comment it out. Need to get a clean way.

Signed-off-by: Yi Liu <yi.l.liu@intel.com>