iommu: Add nested domain support
Introduce a new domain type for a user space I/O address, which is nested
on top of another user space address represented by a UNMANAGED domain. The
mappings of a nested domain are managed by user space software, therefore
it's unnecessary to have map/unmap callbacks. But the updates of the PTEs
in the nested domain page table must be propagated to the caches on both
IOMMU (IOTLB) and devices (DevTLB). This introduces iotlb_sync_user domain
ops for this purpose.

The nested domain is allocated and freed by calling:
    iommu_domain_alloc_user()
    iommu_domain_free();
and set to device or a pasid of device through generic iommu attach/detach
interfaces. The caches are synced by calling:
    iommu_iotlb_sync_user()

The iommu_domain_alloc_user() is extended to allocate a stage-2 user space
domain as well, so that the iommufd could allocate a user domain through a
single interface.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>