vfio: Make open_device exclusive between group path and device cdev path
This prevents userspace from opening a specific vfio_device in both
group path and device cdev path. It is achieved by adding a flag
in vfio_device, in the device cdev path, it is set to prevent another
open including device cdev path and group path. In group path, this
flag is never set.

Signed-off-by: Yi Liu <yi.l.liu@intel.com>