iommufd/device: Move IOAS attaching and detaching operations into helpers
When using a 2-stage nested hwpt setup, the IOVA mappings or SW MSI are at
the stage-2 hwpt, while the device will be attached to a stage-1 hwpt. And
then the current do_attach() flow will fail to set up MSI, since it always
uses the attaching hwpt pointer (i.e. stage-1 in such a case).

As a preparatory change for the following IOMMU_HWPT_ALLOC support patch,
move all related operations into a pair of separate functions, so we only
need to redirect the hwpt pointer to its stage-2 hwpt, when using a nested
setup.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>