iommu: Add nested domain support
Introduce a new domain type for a user I/O page table, which is nested
on top of another user space address represented by a UNMANAGED domain.
The mappings of a nested domain are managed by user space software, so
it is not necessary to have map/unmap callbacks. But the updates of the
PTEs in the nested domain page table must be propagated to the caches on
both IOMMU (IOTLB) and devices (DevTLB).

The nested domain is allocated by the domain_alloc_user op, and attached
to a device through the existing iommu_attach_device/group() interfaces.

New domain ops cache_invalidate_user is added for the userspace to flush
hardware caches for a nested domain through iommufd. No wrapper for it,
as it's only supposed to be used by iommufd.

The invalidation requests being passed in to this cache_invalidate_user
op is in the form of a user data array that conatins a number of entries
to store a list of invalidation data. Add an iommu_user_data_array struct
and its helper function iommu_copy_user_data_from_array for iommu drivers
to walk through the invalidation request array and fetch the data inside.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>
Reviewed-by: Kevin Tian <kevin.tian@intel.com>
Co-developed-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>