iommu: Add nested domain support
Introduce a new domain type for a user space I/O address, which is nested
on top of another user space address represented by a UNMANAGED domain. The
mappings of a nested domain are managed by user space software, therefore
it's unnecessary to have map/unmap callbacks. But the updates of the PTEs
in the nested domain page table must be propagated to the caches on both
IOMMU (IOTLB) and devices (DevTLB).

The nested domain is allocated and freed by calling:
    iommu_domain_alloc_user()
    iommu_domain_free();
and attached to the device through the existing iommu_attach_device/group()
interfaces.

An new domain op, named iotlb_sync_user, and a new iommu interface are
added for the userspace to flush the hardware caches for a nested domain
through iommufd.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>