iommu: Introduce a new iommu_group_replace_domain() API
qemu has a need to replace the translations associated with a domain
when the guest does large scale operations like switching between an
IDENTITY domain and, say, dma-iommu.c.

Currently it does this by replacing all the mappings in a single
domain, but this is very inefficient and means that domains have to be
per-device rather than per-translation.

Provide a high level API to allow replacement of one domain with
another. This is similar to an detach/attach cycle except it doesn't
force the group to go to the blocking domain in-between.

By removing this forced blocking domain the iommu driver has the
opportunity to implement atomic replacement of the domains to the
greatest extent it's hardware allows.

Atomic replacement allows the qemu emulation of the viommu to be more
complete, as real hardware has this ability.

This becomes even more useful when we get into nesting, where there is
value in continuously maintaining the same S2 attached to the device
during all the possible domain transitions. Some scenarios on ARM may
require this to ensure that MSI interrupts are always deliverable.

All drivers are already required to support changing between active
UNMANAGED domains when using their attach_dev ops.

Suggested-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>