iommu: Introduce a new iommu_replace_group() API
For a 2-stage translation case, a group/device can freely switch between
a pair of linked stage-1 and stage-2 domains without being stagged at a
blocking_domain. This can also apply to two domains that have the same
IOVA mappings, since there would be no impact on DMA to the IOVA that it
is accessing.

Add a new iommu_replace_group() for the use cases above. Callers can use
this API as iommu_attach_group() for use cases where the group/device is
attached to a core domain (default_domain or blocking_domain).

Note that drivers should implement default_domain for their platforms to
use this API.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>