iommufd: Add user hw_pagetable allocation
hw_pagetable represents the pagetables used by userspace. Hence it is
meaningful to allow userspace to request hw_pagetable allocation. This
is a mandatory requirement for nested translation support as the stage-1
pagetable is owned by userspace. It is also helpful for replacing the
existing implicit hw_pagetable allocation in iommufd.

hw_pagetables can be stage-1 pagetable, stage-2 pagetable or just standalone
pagetable. For stage-2 pagetable and standalone pagetable, userspace need
to provide an IOAS ID to link the hw_pagetable to an IOAS. For stage-1
pagetable, userspace should provide an ID to point to a stage-2 hw_pagetable
as stage-1 pagetable need to work with stage-2 hw_pagetable.

This adds two ioctls:
    1) IOMMU_HWPT_ALLOC - allocate hw_pagetable
    2) IOMMU_HWPT_INVALIDATE - invalidate stage-1 TLB, this is required by
                               the stage-1 hw_pagetable as it is maintained
			       by userspace.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>