iommufd: Add device_users to track the hw_pagetable usage by device
Currently, hw_pagetable tracks the attached devices using a device list.
When attaching the first device to the kernel-managed hw_pagetable, it
should be linked to IOAS. When detaching the last device from this hwpt,
the link with IOAS should be removed too. And this first-or-last device
check is done with list_empty(hwpt->devices).

Removing the iommufd_hw_pagetable_has_group hack requires a reduction in
the dependency on the device list. As an alternative solution, add a new
refcount device_users in struct iommufd_hw_pagetable to track the users
of a hw_pagetable by the attached devices.

Signed-off-by: Yi Liu <yi.l.liu@intel.com>
Co-developed-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>